potential fix for #9

+++ tiny_agent/agent.py
@@ -89,8 +89,8 @@ class Agent:
         self._links.add(monitored)
         monitored._monitors.add(self)
 
-    def spawn_linked_task(self, coro_or_future) -> 'LinkedTask':
-        return LinkedTask(self, coro_or_future, loop=self._loop)
+    def spawn_linked_task(self, coro_or_future, unlink_on_success: bool = True) -> 'LinkedTask':
+        return LinkedTask(self, coro_or_future, unlink_on_success=unlink_on_success, loop=self._loop)
 
     async def _main(self):
         # noinspection PyBroadException
@@ -131,16 +131,19 @@ class Agent:
 
 
 class LinkedTask(Agent):
-    def __init__(self, linked_to: Agent, coro_or_future, loop: AbstractEventLoop=None):
+    def __init__(self, linked_to: Agent, coro_or_future, unlink_on_success: bool = True,
+                 loop: AbstractEventLoop = None):
         super(LinkedTask, self).__init__(loop=loop)
         self.bidirectional_link(linked_to)
+
         async def do_it():
             try:
                 result = await coro_or_future
                 if result is not None:
                     logger.warning("%s tried to return a result from a LinkedTask, %s", coro_or_future, self)
-                self.unlink(linked_to)  # We finished successfully so lets not kill our friend when we die
+                if unlink_on_success:
+                    self.unlink(linked_to)  # We finished successfully so lets not kill our friend when we die
             finally:
                 self.exit()
-        self._queue.put_nowait(do_it)
 
+        self._queue.put_nowait(do_it)
